+++ xmake/platform/none.lua
@@ -1,7 +1,7 @@
 
 function none_target()
 
-    local use_llvm_toolchain = get_config("use-llvm")
+    local use_llvm_toolchain = get_config("use-llvm-compiler")
     if use_llvm_toolchain then	
         set_toolchains("clang")
         add_ldflags("-fuse-ld=lld", {force = true})