+++ xmake/platform/windows.lua
@@ -3,7 +3,7 @@
 
 function windows_target()
 
-    local use_llvm_toolchain = get_config("use-llvm")
+    local use_llvm_toolchain = get_config("use-llvm-compiler")
     if use_llvm_toolchain then	
         set_toolchains("clang")